Compare Index Event Breakdown between Target and Comparator Cohorts

This function compares the breakdown of index events between a target cohort and a comparator cohort. It calculates the proportions of persons or records (depending on the compareEntries parameter) for each cohort and combines the results into a single data frame. The function also computes standardized differences between the target and comparator cohorts.

compareIndexEventBreakdown(
  indexEventBreakdown,
  cohortCount,
  targetCohortId,
  targetDatabaseId = NULL,
  comparatorCohortId,
  comparatorDatabaseId = targetDatabaseId,
  compareEntries = FALSE
)

Arguments

indexEventBreakdown

A data frame containing index event breakdown information. It should include fields such as cohortDefinitionId, persons, and records, along with databaseId and other event-related details.

cohortCount

A data frame containing cohort counts, including cohortSubjects and cohortEntries for each cohortDefinitionId.

targetCohortId

An integer representing the cohort definition ID of the target cohort to be compared.

targetDatabaseId

(Optional) A string representing the ID of the target database. If NULL, it will be derived from the indexEventBreakdown.

comparatorCohortId

An integer representing the cohort definition ID of the comparator cohort to be compared.

comparatorDatabaseId

(Optional) A string representing the ID of the comparator database. If NULL, it will default to targetDatabaseId.

compareEntries

A boolean indicating whether to compare records (if TRUE) or persons (if FALSE). Defaults to FALSE.

Value

A data frame with the combined breakdown of index events for the target and comparator cohorts. It includes fields such as cohort IDs, database IDs, event concepts, event counts, and computed means (proportions) for both target and comparator cohorts. The function also returns standardized differences between the target and comparator cohorts.

Details

The compareEntries parameter determines whether the comparison is based on the number of records or persons. If compareEntries = TRUE, the function compares records (from cohortEntries and records); otherwise, it compares persons (from cohortSubjects and persons).

The computeStandardizedDifference function is applied to the proportions of both cohorts to calculate the standardized differences.