Get Index Event Breakdown

This function retrieves a breakdown of events (e.g., visits, procedures, drugs, observations, conditions, measurements) associated with the index event (on first cohort start date of a subject) for a given set of cohorts.

getIndexEventBreakdown(
  cohortIds,
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  databaseId,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)

Arguments

cohortIds

A vector of cohort IDs to filter the cohort table for inclusion.

connectionDetails

An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

connection

An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

cdmDatabaseSchema

Schema name where your CDM tables reside.

cohortDatabaseSchema

Schema name where your cohort tables reside. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

cohortTable

The name of the cohort table.

databaseId

A short string for identifying the database (e.g. 'Synpuf').

tempEmulationSchema

Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created.

Value

A tibble containing the breakdown of counts for various clinical events. The tibble consists of:

conceptId

The concept ID of the event (e.g., visit, procedure, drug).

persons

The number of distinct persons having the event.

records

The total number of records of the event.

source

The source of the event count, e.g., 'v1' for visit, 'v2' for visit source, 'p1' for procedure, etc.

Details

For each event type (visit, procedure, drug exposure, observation, condition, and measurement), both standard concept IDs and source concept IDs are retrieved, ensuring no duplication.

This function processes clinical event data for cohorts, such as visits, procedures, drugs, observations, conditions, and measurements, based on the index date (cohort start date). It counts the number of persons and records for each event and combines them into a single tibble, distinguishing between standard and source concepts.

SQL queries are translated and executed using the DatabaseConnector package. Temporary tables are dropped after processing.