Get Predicted Count in a Data Frame

This function predicts the count of events in a data frame using Poisson regression models with natural splines. It fits the model using both the Cyclops package and base glm(). The function models event counts based on the field specified in the timeSequenceField parameter and includes an offset for the field specified in personTimeField to account for different exposure times. It evaluates model stability through likelihood ratio tests and other diagnostics. The function adds flags to indicate model stability and significance, with separate indicators for the results of both Cyclops and glm() fits.

getPredictedEventCounts(
  data,
  timeSequenceField,
  countField,
  personTimeField,
  maxNumberOfSplines = NULL,
  splineTickInterval = 3,
  alpha = 0.05,
  maxRatio = 1.25
)

Arguments

data

A data frame that contains the columns specified in the function parameters.

timeSequenceField

The field in the data frame that represents the time sequence.

countField

The field in the data frame that represents the count of events.

personTimeField

The field in the data frame that represents the person time.

maxNumberOfSplines

The maximum number of splines to use in the model. If NULL, the default is 3.

splineTickInterval

The interval at which to place the splines. The default is 3.

alpha

The significance level for statistical tests and confidence intervals (default is 0.05).

maxRatio

The maximum ratio for the likelihood comparison. If NULL, the default is 1.25.

Value

A data frame with the observed and predicted counts from both Cyclops and glm(), along with diagnostics such as p-values, confidence intervals (for GLM), and stability indicators.

Details

Input Data Structure: The input data should be a data frame with columns that are dynamically specified via function parameters.

The timeSequenceField parameter represents a column containing the time sequence or index of the events (e.g., date or time intervals).

The countField parameter refers to the column representing the observed count of events during each time interval.

The personTimeField parameter points to the column representing the exposure time or population-at-risk during the corresponding time interval. The data should not contain duplicate rows for the same value in the column specified by timeSequenceField, and both the columns for countField and personTimeField must not contain missing (NA) values.